+++ tests/agreement_test.rs
@@ -14,8 +14,8 @@ fn test_create_agreement() {
     let mut agent = load_test_agent_one();
     let mut agent_two = load_test_agent_two();
     let mut agentids: Vec<String> = Vec::new();
-    agentids.push(agent.get_lookup_id().expect("REASON"));
-    agentids.push(agent_two.get_lookup_id().expect("REASON"));
+    agentids.push(agent.get_id().expect("REASON"));
+    agentids.push(agent_two.get_id().expect("REASON"));
 
     let document_string =
         load_local_document(&"examples/documents/e957d062-d684-456b-8680-14a1c4edcb2a:5599ac70-a3d6-429b-85ae-c9b17c78d2c5.json".to_string()).unwrap();
@@ -25,6 +25,8 @@ fn test_create_agreement() {
     let unsigned_doc = agent
         .create_agreement(&document_key, &agentids)
         .expect("create_agreement");
+
+    println!("{}", unsigned_doc.to_string());
 
     // agent one  tries and fails to creates agreement document
     let result = agent.create_agreement(&document_key, &agentids);